Stekim4/fix review cards

+++ server/controllers/bikeTrailsController.js
@@ -9,8 +9,8 @@ bikeController.getTrails = async (req, res, next) => {
         const options = {
             method: 'GET',
             headers: {
-              'X-RapidAPI-Key': '8e8ae2bc71mshfbc5c59d3dfd9afp1b60cejsnad30f37fbe25',
-              'X-RapidAPI-Host': 'trailapi-trailapi.p.rapidapi.com'
+              'X-RapidAPI-Key': process.env.VITE_TRAILAPI_KEY,
+              'X-RapidAPI-Host': process.env.VITE_TRAILAPI_HOST
             }
           };
         const trailsAPIResponse = await fetch(url, options);
@@ -34,16 +34,13 @@ bikeController.getTrails = async (req, res, next) => {
 
 bikeController.getFavTrails = async (req, res, next) => {
   try {
-    const user_id = req.cookies.userID;
-    const getTrailsSQL = `
-    SELECT * FROM favorite_trails
-    WHERE google_id = '${user_id}';`;
-
-    let dbRes = await db.query(getTrailsSQL);
+    const user_id = req.user.user_id;
+    const params = [user_id]
+    const getTrailsSQL = `SELECT * FROM favorite_trails WHERE user_id = $1;`;
+    let dbRes = await db.query(getTrailsSQL, params);
     const trailsForQuery = [];
     for (let i = 0; i < dbRes.rows.length; i++) {
       const trail = dbRes.rows[i];
-
       trailsForQuery.push({ trailId: trail['trail_id'], trailName: trail['trail_name']});
     }
 
@@ -58,23 +55,23 @@ bikeController.getFavTrails = async (req, res, next) => {
 
 bikeController.saveTrails = async (req, res, next) => {
   try {
-    const google_id = req.cookies.userID;
+    // const google_id = req.cookies.userID;
+    const user_id = req.user.user_id;
     const { id } = req.body;
-    const checkTrailsSQL = `
-    SELECT * FROM favorite_trails
-    WHERE google_id = '${google_id}' AND trail_id = '${id}';`;
+    const params = [user_id, id];
+    const checkTrailsSQL = `SELECT * FROM favorite_trails WHERE user_id = $1 AND trail_id = $2;`;
     // query first to db to check if trail is already favorited to prevent duplicates then save trail
-    let alreadyFavorited = await db.query(checkTrailsSQL)
+    let alreadyFavorited = await db.query(checkTrailsSQL, params)
     alreadyFavorited = await alreadyFavorited.json();
     res.locals.isSaved = true;
     return next();
   } catch(err) {
-    const google_id = req.cookies.userID;
+    const user_id = req.user.user_id;
+    // const google_id = req.cookies.userID;
     const { id, name } = req.body;
-    const saveTrailsSQL = `
-    INSERT INTO favorite_trails(google_id, trail_id, trail_name)
-    VALUES('${google_id}', '${id}', '${name}');`;
-    const saveQuery = await db.query(saveTrailsSQL);
+    const params = [user_id, id, name];
+    const saveTrailsSQL = `INSERT INTO favorite_trails(user_id, trail_id, trail_name) VALUES($1, $2, $3);`;
+    const saveQuery = await db.query(saveTrailsSQL, params);
     res.locals.isSaved = true;
     next();
   }
@@ -83,12 +80,12 @@ bikeController.saveTrails = async (req, res, next) => {
 
 bikeController.deleteTrails = async (req, res, next) => {
   try {
-    const user_id = req.cookies.userID;
+    // const user_id = req.cookies.userID;
+    const user_id = req.user.user_id;
     const { trailId } = req.params;
-    const deleteTrailsSQL = `
-    DELETE FROM favorite_trails
-    WHERE google_id ='${user_id}' AND trail_id = '${trailId}';`
-    await db.query(deleteTrailsSQL)
+    const params = [user_id, trailId]
+    const deleteTrailsSQL = `DELETE FROM favorite_trails WHERE user_id = $1 AND trail_id = $2;`
+    await db.query(deleteTrailsSQL, params)
     res.locals.isDeleted = true;
     return next();
   } catch(err) {
